requirements.txt
streamlit
scikit-learn
pandas
numpy
python-dotenv
google-generativeai